feat(engine): fixed-point int64 pricing (x10000)

+++ internal/dto/orderbook.go
@@ -12,7 +12,7 @@ type OrderbookDepthResponse struct {
 }
 
 type OrderBookEntry struct {
-	Price    float64 `json:"price"`
-	Quantity float64 `json:"quantity"`
-	Total    float64 `json:"total"`
+	Price    int64 `json:"price"`
+	Quantity int64 `json:"quantity"`
+	Total    int64 `json:"total"`
 }
